Calculating a transformation of continuous feature using PDP/ALE plot

The safely_transform_continuous() function calculates a transformation function for the continuous variable using a PD/ALE plot obtained from black box model.

safely_transform_continuous(explainer, variable, response_type = "ale",
  grid_points = 50, N = 200, penalty = "MBIC", nquantiles = 10,
  no_segments = 2)

Arguments

explainer
DALEX explainer created with explain() function

variable
a feature for which the transformation function is to be computed

response_type
character, type of response to be calculated, one of: "pdp", "ale". If features are uncorrelated, one can use "pdp" type - otherwise "ale" is strongly recommended.

grid_points
number of points on x-axis used for creating the PD/ALE plot, default 50

N
number of observations from the dataset used for creating the PD/ALE plot, default 200

penalty
penalty for introducing another changepoint, one of "AIC", "BIC", "SIC", "MBIC", "Hannan-Quinn" or numeric non-negative value

nquantiles
the number of quantiles used in integral approximation

no_segments
numeric, a number of segments variable is to be divided into in case of founding no breakpoints

Value

list of information on the transformation of given variable

See also

Examples

library(DALEX)
library(randomForest)
library(rSAFE)

data <- apartments[1:500,]
set.seed(111)
model_rf <- randomForest(m2.price ~ construction.year + surface + floor +
                           no.rooms + district, data = data)
explainer_rf <- explain(model_rf, data = data[,2:6], y = data[,1])
safely_transform_continuous(explainer_rf, "construction.year")

#> $sv
#>                                                           _vname_      _label_
#> construction.year.randomForest.1920             construction.year randomForest
#> construction.year.randomForest.1923             construction.year randomForest
#> construction.year.randomForest.1924             construction.year randomForest
#> construction.year.randomForest.1926             construction.year randomForest
#> construction.year.randomForest.1927             construction.year randomForest
#> construction.year.randomForest.1929             construction.year randomForest
#> construction.year.randomForest.1931             construction.year randomForest
#> construction.year.randomForest.1933             construction.year randomForest
#> construction.year.randomForest.1935             construction.year randomForest
#> construction.year.randomForest.1936             construction.year randomForest
#> construction.year.randomForest.1938             construction.year randomForest
#> construction.year.randomForest.1940             construction.year randomForest
#> construction.year.randomForest.1943.20408163265 construction.year randomForest
#> construction.year.randomForest.1945             construction.year randomForest
#> construction.year.randomForest.1947             construction.year randomForest
#> construction.year.randomForest.1948.75510204082 construction.year randomForest
#> construction.year.randomForest.1950             construction.year randomForest
#> construction.year.randomForest.1951             construction.year randomForest
#> construction.year.randomForest.1953             construction.year randomForest
#> construction.year.randomForest.1956             construction.year randomForest
#> construction.year.randomForest.1958             construction.year randomForest
#> construction.year.randomForest.1960             construction.year randomForest
#> construction.year.randomForest.1962             construction.year randomForest
#> construction.year.randomForest.1964             construction.year randomForest
#> construction.year.randomForest.1966.59183673469 construction.year randomForest
#> construction.year.randomForest.1968             construction.year randomForest
#> construction.year.randomForest.1970             construction.year randomForest
#> construction.year.randomForest.1971             construction.year randomForest
#> construction.year.randomForest.1973             construction.year randomForest
#> construction.year.randomForest.1975             construction.year randomForest
#> construction.year.randomForest.1977             construction.year randomForest
#> construction.year.randomForest.1979.87755102041 construction.year randomForest
#> construction.year.randomForest.1983             construction.year randomForest
#> construction.year.randomForest.1985             construction.year randomForest
#> construction.year.randomForest.1986             construction.year randomForest
#> construction.year.randomForest.1987             construction.year randomForest
#> construction.year.randomForest.1989             construction.year randomForest
#> construction.year.randomForest.1990             construction.year randomForest
#> construction.year.randomForest.1992             construction.year randomForest
#> construction.year.randomForest.1993             construction.year randomForest
#> construction.year.randomForest.1994             construction.year randomForest
#> construction.year.randomForest.1996             construction.year randomForest
#> construction.year.randomForest.1997             construction.year randomForest
#> construction.year.randomForest.1999             construction.year randomForest
#> construction.year.randomForest.2001             construction.year randomForest
#> construction.year.randomForest.2003.44897959184 construction.year randomForest
#> construction.year.randomForest.2005             construction.year randomForest
#> construction.year.randomForest.2007             construction.year randomForest
#> construction.year.randomForest.2010             construction.year randomForest
#>                                                      _x_     _yhat_ _ids_
#> construction.year.randomForest.1920             1920.000    0.00000     0
#> construction.year.randomForest.1923             1923.000   24.46394     0
#> construction.year.randomForest.1924             1924.000   19.25767     0
#> construction.year.randomForest.1926             1926.000   11.36508     0
#> construction.year.randomForest.1927             1927.000  -20.25104     0
#> construction.year.randomForest.1929             1929.000  -36.89223     0
#> construction.year.randomForest.1931             1931.000  -61.07567     0
#> construction.year.randomForest.1933             1933.000  -93.95953     0
#> construction.year.randomForest.1935             1935.000 -226.91019     0
#> construction.year.randomForest.1936             1936.000 -265.47288     0
#> construction.year.randomForest.1938             1938.000 -286.52685     0
#> construction.year.randomForest.1940             1940.000 -311.84111     0
#> construction.year.randomForest.1943.20408163265 1943.204 -316.86066     0
#> construction.year.randomForest.1945             1945.000 -333.69903     0
#> construction.year.randomForest.1947             1947.000 -337.82020     0
#> construction.year.randomForest.1948.75510204082 1948.755 -350.60562     0
#> construction.year.randomForest.1950             1950.000 -358.25608     0
#> construction.year.randomForest.1951             1951.000 -366.40703     0
#> construction.year.randomForest.1953             1953.000 -363.90753     0
#> construction.year.randomForest.1956             1956.000 -369.63511     0
#> construction.year.randomForest.1958             1958.000 -365.59776     0
#> construction.year.randomForest.1960             1960.000 -360.55766     0
#> construction.year.randomForest.1962             1962.000 -354.60478     0
#> construction.year.randomForest.1964             1964.000 -370.41073     0
#> construction.year.randomForest.1966.59183673469 1966.592 -375.87497     0
#> construction.year.randomForest.1968             1968.000 -376.18428     0
#> construction.year.randomForest.1970             1970.000 -372.00503     0
#> construction.year.randomForest.1971             1971.000 -377.12050     0
#> construction.year.randomForest.1973             1973.000 -373.42861     0
#> construction.year.randomForest.1975             1975.000 -378.26315     0
#> construction.year.randomForest.1977             1977.000 -380.34691     0
#> construction.year.randomForest.1979.87755102041 1979.878 -374.19200     0
#> construction.year.randomForest.1983             1983.000 -370.07926     0
#> construction.year.randomForest.1985             1985.000 -364.84144     0
#> construction.year.randomForest.1986             1986.000 -362.46225     0
#> construction.year.randomForest.1987             1987.000 -363.21756     0
#> construction.year.randomForest.1989             1989.000 -350.81047     0
#> construction.year.randomForest.1990             1990.000 -347.50571     0
#> construction.year.randomForest.1992             1992.000 -350.08075     0
#> construction.year.randomForest.1993             1993.000 -340.35548     0
#> construction.year.randomForest.1994             1994.000 -323.15982     0
#> construction.year.randomForest.1996             1996.000  -85.53955     0
#> construction.year.randomForest.1997             1997.000  -85.57973     0
#> construction.year.randomForest.1999             1999.000  -78.09553     0
#> construction.year.randomForest.2001             2001.000  -64.20390     0
#> construction.year.randomForest.2003.44897959184 2003.449  -59.82099     0
#> construction.year.randomForest.2005             2005.000  -27.68647     0
#> construction.year.randomForest.2007             2007.000  -14.82230     0
#> construction.year.randomForest.2010             2010.000  -15.52523     0
#> 
#> $break_points
#> [1] 1926.000 1940.000 1962.000 1979.878 1994.000
#> 
#> $new_levels
#> [1] "(-Inf, 1926]"     "(1926, 1940]"     "(1940, 1962]"     "(1962, 1979.878]"
#> [5] "(1979.878, 1994]" "(1994, Inf)"     
#>